Remove SeatGroup.__add__() and add SeatGroup.math_operation(), .subtract().  Add copy function to SeatGroup.get_seats*()
SeatGroup.__add__() removed as it was not so intuitive that it should be a merge.  At first __add__ was treated as a merger like lists, but SeatGroup feels more like a np.array sort of object so the __add__() probably fits better like np.array's add.  To be replaced later with binding to SG.math_operation('add'), etc.